+++ reporte/src/main/java/Reporte/reporte/model/UsuarioDTO.java
@@ -10,8 +10,8 @@ public class UsuarioDTO {
         return id;
     }
 
-    public void setId(int id) {
-        this.id = (long) id;
+    public void setId(Long id) {
+        this.id = id;
     }
 
     public String getNombres() {
